+++ .Tools/CI/Settings/ProBuilderSettings.cs
@@ -55,8 +55,8 @@ public class ProBuilderSettings : AnnotatedSettingsBase
         Wrench.Packages[ProBuilderPackageName].CoverageCommands.Enabled = true;
         Wrench.Packages[ProBuilderPackageName].CoverageCommands.Commands = [_excludeAssembliesCodeCovCommand];
 
-        var defaultMacPlatform = WrenchPackage.DefaultEditorPlatforms[SystemType.MacOS];
-        Wrench.Packages["com.unity.probuilder"].EditorPlatforms[SystemType.MacOS] = new Platform(new Agent("package-ci/macos-13-arm64:v4", FlavorType.MacDefault, defaultMacPlatform.Agent.Resource, "M1"), defaultMacPlatform.System);
+        // var defaultMacPlatform = WrenchPackage.DefaultEditorPlatforms[SystemType.MacOS];
+        // Wrench.Packages["com.unity.probuilder"].EditorPlatforms[SystemType.MacOS] = new Platform(new Agent("package-ci/macos-13-arm64:v4", FlavorType.MacDefault, defaultMacPlatform.Agent.Resource, "M1"), defaultMacPlatform.System);
     }
 
     public WrenchSettings Wrench { get; private set; }